refactor: Remove config.py and refactor structure
