+++ docker-compose.yml
@@ -3,33 +3,33 @@ version: '3.8'
 services:
   db:
     image: mariadb:latest
-    container_name: hashfish_mariadb2
+    container_name: hashfish_mariadb
     environment:
-      MYSQL_DATABASE: ${dbname}
-      MYSQL_ROOT_PASSWORD: ${dbpassword}
+      MYSQL_DATABASE: ${DB_NAME}
+      MYSQL_ROOT_PASSWORD: ${DB_PASSWORD}
     ports:
       - "3306:3306"
     volumes:
       - db_data:/var/lib/mysql
 
   app:
-    image: harksulim/hashfish:2
-    container_name: hashfish2
+    image: harksulim/hashfish:latest
+    container_name: hashfish
     environment:
-      DB_HOST: ${dbhost}
+      DB_HOST: ${DB_HOST}
       DB_PORT: 3306
-      DB_USER: ${dbuser}
-      DB_PASSWORD: ${dbpassword}
-      DB_NAME: ${dbname}
-      MASTER_EMAIL: ${master_email}
-      MASTER_APIKEY: ${master_apikey}
+      DB_USER: ${DB_USER}
+      DB_PASSWORD: ${DB_PASSWORD}
+      DB_NAME: ${DB_NAME}
+      MASTER_EMAIL: ${MASTER_EMAIL}
+      MASTER_APIKEY: ${MASTER_APIKEY}
     ports:
       - "8080:8080"
     depends_on:
       - db
     volumes:
       - .:/app
-    command: sh -c "sleep 10 && /app/start.sh"
+    command: sh -c "sleep 5 && /app/start.sh"
 
 volumes:
   db_data:
