+++ requirements.txt
@@ -1,7 +1,8 @@
-fastapi==0.95.1
-uvicorn==0.22.0
-SQLAlchemy==2.0.9
-requests==2.28.2
-PyMySQL==1.0.3
+fastapi==0.112.0
+uvicorn==0.30.5
+pydantic==2.8.2
+SQLAlchemy==2.0.32
+PyMySQL==1.1.1
 python-dotenv==1.0.1
+requests==2.32.3
 #파이썬 버전도?
