+++ start.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 echo "RUN API"
-python3 main.py
+python3 fish-hash/main.py
